Extracted Point class
Extracted the inner class Point from Map and removed Point’s dependency
on Map.
