+++ .gitignore
@@ -6,3 +6,4 @@
 /Map.class
 /SearchPoint.class
 /AStarSearch.class
+/Point.class
